src/requirements.txt
discord.py
request
python-dotenv